Support module paths outside of Magento root
In my Magento environment, I'm using local composer repositories that are outside of the Magento root. In other words, if Magento lives in `/var/www/html`, a Magento module could actually be living in `/data/extensions/`. Unfortunately, this JS util is expecting the module path to be part of the Magento root, and because of that, it fails to detect a custom Tailwind file. This PR adds support for module paths outside of Magento, simply by seeing if the delivered (absolute) path exists.